clean and rename mycoaddcam.py to coaddcam.py
